Make supersedes for addable, refactor existng moveable code
