+++ assets/src/pages/formset.js
@@ -1,0 +1,5 @@
+import init from 'src/lib/init.js'
+import initFormset from 'src/lib/formset.js'
+
+init()
+initFormset()
