+++ python/aristotle-metadata-registry/aristotle_mdr/contrib/reviews/templates/aristotle_mdr/reviews/review/supersedes.html
@@ -1,5 +1,14 @@
 {% extends "aristotle_mdr/reviews/review/base.html" %}
 {% load util_tags %}
+{% load render_bundle from webpack_loader %}
+
+{% block webpack_bundle %}
+    {% render_bundle 'formset' 'js' %}
+{% endblock %}
+
+{% block webpack_css_bundle %}
+    {% render_bundle 'formset' 'css' %}
+{% endblock %}
 
 {% block title %}Review details{{ item.name }}{% endblock %}
 {% block tabpanel %}
@@ -7,7 +16,7 @@
   <div class="panel">
     <div class="panel-body">
       These supersedes relations are being proposed as part of this request
-      <form method="post">
+      <form id="{{ formset.prefix }}" method="post">
         {% csrf_token %}
         {{ formset.management_form }}
         {% for form in formset %}
@@ -15,12 +24,6 @@
             {% for field in form.hidden_fields %}
               {{ field }}
             {% endfor %}
-            {% if form.initial.newer_item %}
-              <div class="form-group">
-                <label>Newer item:</label>
-                <span>{{ names|lookup_string:form.initial.newer_item }}</span>
-              </div>
-            {% endif %}
             {% for field in form.visible_fields %}
               <div class="form-group">
                 {{ field.errors }}
@@ -30,8 +33,26 @@
             {% endfor %}
           </div>
         {% endfor %}
+        <a class="add_code_button btn btn-success" formid="{{ formset.prefix }}" href="#">Add</a>
         <button type="submit" class="btn btn-primary">Submit</button>
       </form>
+
+      <div id="{{ formset.prefix }}" class="formstage" style="display:none">
+        {% with formset.empty_form as form %}
+          <div class="form-inline">
+            {% for field in form.hidden_fields %}
+              {{ field }}
+            {% endfor %}
+            {% for field in form.visible_fields %}
+              <div class="form-group">
+                {{ field.label_tag }}
+                {{ field }}
+              </div>
+            {% endfor %}
+          </div>
+        {% endwith %}
+      </div>
+
     </div>
   </div>
 </div>
